Split out xdg specific code to own file
