+++ compositor/xdg.h
@@ -1,0 +1,8 @@
+#ifndef XDG_H
+#define XDG_H
+
+#include "server.h"
+
+void wc_init_xdg(struct wc_server* server);
+
+#endif//XDG_H
